load balancing: moved DumbPolicy to mod
This way it will be cleaner to use it for testing of various load
balancing policies.
